pyproject.toml
[build-system]
requires = [
    "hatchling==1.27.0",
    "hatch-vcs==0.5.0",
    "cloudpickle==3.1.1",
    "pyzmq==27.0.1",
]
build-backend = "hatchling.build"

[project]
name = "executorlib"
description = "Up-scale python functions for high performance computing (HPC) with executorlib."
authors = [
    { name = "Jan Janssen", email = "janssen@lanl.gov" },
]
readme = "README.md"
license = { file = "LICENSE" }
keywords = ["high performance computing", "hpc", "task scheduler", "slurm", "flux-framework", "executor"]
requires-python = ">3.9, <3.14"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Topic :: Scientific/Engineering :: Physics",
    "License :: OSI Approved :: BSD License",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "cloudpickle==3.1.1",
    "pyzmq==27.0.1",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/pyiron/executorlib"
Documentation = "https://executorlib.readthedocs.io"
Repository = "https://github.com/pyiron/executorlib"

[project.optional-dependencies]
cache = ["h5py==3.14.0"]
graph = [
    "pygraphviz==1.14",
    "networkx==3.4.2",
]
graphnotebook = [
    "pygraphviz==1.14",
    "networkx==3.4.2",
    "ipython==9.0.2",
]
mpi = ["mpi4py==4.0.1"]
cluster = [
    "pysqa==0.3.0",
    "h5py==3.14.0",
]
all = [
    "mpi4py==4.0.1",
    "pysqa==0.3.0",
    "h5py==3.14.0",
    "pygraphviz==1.14",
    "networkx==3.4.2",
    "ipython==9.0.2",
]

[tool.ruff]
exclude = [".ci_support", "docs", "notebooks", "tests", "setup.py", "_version.py"]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # flake8-comprehensions
    "C4",
    # eradicate
    "ERA",
    # pylint
    "PL",
]
ignore = [
    # ignore line-length violations
    "E501",
    # ignore imports which are not at the top of a module
    "PLC0415",
    # Too many arguments in function definition
    "PLR0913",
    # Magic value used in comparison
    "PLR2004",
    # Too many branches
    "PLR0912",
]

[tool.hatch.build]
include = [
    "executorlib"
]

[tool.hatch.build.hooks.vcs]
version-file = "executorlib/_version.py"

[tool.hatch.build.targets.sdist]
include = [
    "executorlib"
]

[tool.hatch.build.targets.wheel]
packages = [
    "executorlib"
]

[tool.hatch.version]
source = "vcs"
path = "executorlib/_version.py"

[tool.coverage.run]
omit = ["executorlib/_version.py", "tests/*"]
command_line = "-m unittest discover tests"